requirements.txt
Flask>=2.3
google-generativeai>=0.5.0
python-dotenv>=1.0.1
langchain>=0.1.16
langchain-core>=0.1.35
langchain_google_genai>=0.0.10
langchain_community>=0.0.30
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
chromadb>=0.4.24
sentence_transformers
flask-cors
groq
